refactor: 검색 페이지에서 검색 기능 커스텀 훅 분리 (#5)
